Polished the app (add button, list errors, etc)

+++ lib/models/MovieModels/action_movies_model.dart
@@ -26,6 +26,6 @@ class ActionMovieModel extends MovieModel {
         release_date: json['release_date'],
         // run_time: json['runtime'],
         // genre: json['genres'],
-        rating: json['vote_average']);
+        rating: json['vote_average'].toDouble());
   }
 }
